feat: ユーザー登録時のイベントハンドラをhooksディレクトリにて宣言
