+++ src/app/user/register/page.tsx
@@ -9,16 +9,13 @@ import Information from "@/components/user/Information/Information";
 import { registerSchema } from "@/lib/validation";
 import type { FormData } from "@/types/user";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useRouter } from "next/navigation";
-import { useState } from "react";
 import { type SubmitHandler, useForm } from "react-hook-form";
-import { ToastContainer, toast } from "react-toastify";
+import { ToastContainer } from "react-toastify";
 import styles from "./page.module.css";
 import "react-toastify/dist/ReactToastify.css";
+import { userRegister } from "@/hooks/useRegister";
 
 const UserRegistration = () => {
-  // useStateでサーバーエラー管理
-  const [serverError, setServerError] = useState<string | null>(null);
   // React hook formでフォーム管理
   const {
     register,
@@ -28,44 +25,9 @@ const UserRegistration = () => {
     resolver: zodResolver(registerSchema),
   });
 
-  const router = useRouter();
+  const { RegisterUser,serverError } = userRegister();
 
-  const onSubmit: SubmitHandler<FormData> = async (data: FormData) => {
-    try {
-      const response = await fetch("/api/user/register", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          name: data.name,
-          email: data.email,
-          password: data.password,
-        }),
-      });
-
-      if (!response.ok) {
-        // 詳細なエラーメッセージ取得
-        const error = await response.json();
-        setServerError(error.message);
-      } else {
-        toast.success("アカウント登録が完了しました！", {
-          position: "top-center",
-          autoClose: 1000,
-          closeButton: true,
-          hideProgressBar: true,
-          closeOnClick: true,
-          theme: "colored",
-        });
-        setTimeout(() => {
-          router.push("/user/login");
-        }, 1500);
-      }
-    } catch (err) {
-      console.log(err);
-      setServerError("予期せぬエラーが発生しました");
-    }
-  };
+  const onSubmit = (data:FormData) => {RegisterUser(data)};
 
   return (
     <>
